python-control-layer/requirements.txt
paho-mqtt>=1.6.1
protobuf>=4.25.8  # Security: CVE fixes for DoS vulnerabilities
numpy>=1.24.0
python-dotenv>=1.0.0
fastapi>=0.109.1  # Security: Fix for ReDoS vulnerability
uvicorn>=0.24.0
pydantic>=2.4.0
psycopg2-binary>=2.9.9
websockets>=12.0
hvac>=2.1.0  # HashiCorp Vault client (optional)
